git add

+++ src/ParkingLot/models/Operator.java
@@ -2,7 +2,7 @@ package ParkingLot.models;
 
 public class Operator extends BaseModel{
     private String name;
-    private int empId;
+    private Long empId;
 
     public String getName() {
         return name;
@@ -12,11 +12,11 @@ public class Operator extends BaseModel{
         this.name = name;
     }
 
-    public int getEmpId() {
+    public Long getEmpId() {
         return empId;
     }
 
-    public void setEmpId(int empId) {
+    public void setEmpId(Long empId) {
         this.empId = empId;
     }
 }
